# api/docs/api-contract.yaml
openapi: 3.1.0
info:
  title: Onboarding API
  version: 1.12.0
  description: |
    Backend API for onboarding durable session storage and ingestion.

    ## Versioning

    This contract follows semantic versioning. The `version` field tracks
    breaking changes (major), new endpoints/fields (minor), and documentation
    fixes (patch). See `docs/api-versioning.md` for the full change policy.

    ## Authentication

    All endpoints under `/api/onboarding` require two headers:
    - `Authorization: Bearer <cognito-access-token>`
    - `X-ID-Token: <cognito-id-token>`

    The ID token carries the `custom:tenant` claim used for tenant isolation.

    ## Error Convention

    Every error response uses a stable shape:
    ```json
    { "error": { "code": "<ERROR_CODE>", "message": "...", "requestId": "..." } }
    ```
    Clients should match on `code` (not `message`) for programmatic handling.
servers:
  - url: /api/onboarding
    description: Onboarding API base path
security:
  - cognitoBearer: []
    cognitoIdToken: []
paths:
...
  /session:
    get:
      operationId: getSession
      summary: Get current onboarding session
      description: |
        Returns the active onboarding session for the authenticated user and tenant.
        If no session exists, returns 404.
      responses:
        "200":
          description: Session data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createSession
      summary: Create a new onboarding session
      description: |
        Creates a new onboarding session for the authenticated user and tenant.
        Returns 409 if an active session already exists.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateSession
      summary: Update the current onboarding session
      description: |
        Merges the provided fields into the active session.
        Only top-level fields in `data` are replaced; omitted fields are preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /uploads/presign:
    post:
      operationId: createImageUploadUrl
      summary: Get a presigned URL for image upload
      description: |
        Returns a presigned S3 PUT URL that the client uses to upload an image directly.
        Object keys are session-scoped under the tenant prefix.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadPresignRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadPresignResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/InternalError"
        "500":
          $ref: "#/components/responses/InternalError"
  /uploads/{key}:
    get:
      operationId: createImageDownloadUrl
      summary: Get a presigned URL for image retrieval
      description: |
        Returns a short-lived presigned S3 GET URL for a tenant-owned upload object.
        The `key` path parameter must be URL-encoded when it contains `/`.
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          description: URL-encoded S3 object key under the tenant upload prefix.
      responses:
        "200":
          description: Presigned download URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadDownloadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/InternalError"
        "500":
          $ref: "#/components/responses/InternalError"
  /gmail/oauth/start:
    post:
      operationId: startGmailOauth
      summary: Start Gmail OAuth
      description: |
        Returns an OAuth authorization URL for the user to connect Gmail.

        Notes:
        - The caller must be authenticated (Cognito headers).
        - The returned URL redirects to `/gmail/oauth/callback` on this service.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GmailOAuthStartRequest"
      responses:
        "200":
          description: OAuth URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailOAuthStartResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /gmail/status:
    get:
      operationId: getGmailStatus
      summary: Gmail connection status
      description: |
        Returns whether Gmail is configured and connected for the current user.
      responses:
        "200":
          description: Status response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /barcode/lookup:
    get:
      operationId: lookupBarcode
      summary: Barcode lookup (query)
      description: |
        Legacy query-string barcode lookup. Returns product metadata when available.
        When providers are temporarily unavailable, returns an unenriched response
        with `resultState=provider_unavailable`.
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: Barcode/GTIN code to look up.
      responses:
        "200":
          description: Barcode lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarcodeLookupResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: lookupBarcodePost
      summary: Barcode lookup
      description: |
        Look up a barcode from a request body payload. Returns an enriched product
        when found, an empty result when not found, or an unenriched result when
        upstream providers are temporarily unavailable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarcodeLookupRequest"
      responses:
        "200":
          description: Barcode lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarcodeLookupResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /urls/scrape:
    post:
      operationId: scrapeUrls
      summary: URL scrape
      description: |
        Scrape and normalize up to 50 URLs and return reviewable item metadata.
        The backend attempts Playwright rendering first, then jsdom parsing, and
        finally a Jina reader fallback when earlier providers fail or produce no
        useful metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UrlScrapeRequest"
      responses:
        "200":
          description: URL scrape response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlScrapeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /photos/analyze:
    post:
      operationId: analyzePhoto
      summary: Photo analysis (Gemini)
      description: |
        Analyze an uploaded onboarding image and return extracted product metadata.
        When `GEMINI_API_KEY` is not configured, the endpoint degrades gracefully
        and returns `analyzed: false` with `reason: gemini_unavailable`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoAnalyzeRequest"
      responses:
        "200":
          description: Photo analysis response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoAnalyzeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions:
    post:
      operationId: createOnboardingSession
      summary: Create a new onboarding session
      description: |
        Creates a new Redis-backed onboarding session for the authenticated user and tenant.
        Returns the session ID plus deep-link URLs for mobile barcode and photo capture.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{sessionId}:
    get:
      operationId: getOnboardingSession
      summary: Read session metadata
      description: |
        Returns metadata for an existing onboarding session. Authenticated users can
        only read sessions they own (tenantId + userId must match). Mobile token flow
        is not supported on this endpoint — Cognito headers are required.
      parameters:
        - $ref: "#/components/parameters/SessionIdPath"
      responses:
        "200":
          description: Session metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionMetaResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /images/upload:
    post:
      operationId: uploadImage
      summary: Server-side image upload to S3
      description: |
        Accepts a base64 data URL, validates the image, and stores it in S3.
        Returns the public or CDN URL for the uploaded image.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadRequest"
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    cognitoBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito access token
    cognitoIdToken:
      type: apiKey
      in: header
      name: X-ID-Token
      description: Cognito ID token (carries custom:tenant claim)
  schemas:
...
    SessionResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Session"
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: Session not found
              requestId: 550e8400-e29b-41d4-a716-446655440000
...